MainView 생성 - 2가지 화면 이동 선택
