+++ RxSwiftPractices/Views/CharactersViewController.swift
@@ -1,0 +1,23 @@
+//
+//  CharactersViewController.swift
+//  RxSwiftPractices
+//
+//  Created by 여나경 on 2021/09/12.
+//
+
+import UIKit
+
+class CharactersViewController: UIViewController {
+
+    // TODO: Make tableView with detailed information
+    // Tableview - without storyboard!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        
+        
+    }
+    
+
+}
